+++ CHANGES.rst
@@ -6,6 +6,10 @@ Changelog
 
 - Nothing changed yet.
 
+- Change pickle format to export the microseconds as an int, to
+  solve a problem with dates after 2038.
+  (`#56 <https://github.com/zopefoundation/DateTime/issues/56>`_)
+
 
 5.4 (2023-12-15)
 ----------------